test: Deduplicate definitions for tests requiring DBus
Ticket: MEN-6656
Changelog: none
